.spice/models/requirements.test.txt
spicepy @ git+https://github.com/spiceai/spicepy@v1.0.0
duckdb==0.9.1
pandas==2.0.3
numpy==1.26.1
pyyaml==6.0.1
web3==6.11.1
joblib==1.3.2
numpy==1.26.1
scikit-learn==1.3.2
scipy==1.11.3
threadpoolctl==3.2.0
tqdm==4.66.1
xgboost==2.0.1